Refactor Store class to a module
